split smath tests
